feat: Remove field mapping functionality and associated fields from Doc2Sys integration settings
