+++ doc2sys/doc2sys/doctype/doc2sys_integration_settings/doc2sys_integration_settings.py
@@ -4,15 +4,6 @@ from doc2sys.integrations.registry import IntegrationRegistry
 
 class Doc2SysIntegrationSettings(Document):
     
-    def before_save(self):
-        """Process field mapping JSON if it's provided as a string"""
-        if isinstance(self.field_mapping, str):
-            try:
-                import json
-                json.loads(self.field_mapping)  # Just validate it's valid JSON
-            except Exception:
-                frappe.throw("Field Mapping must be a valid JSON object")
-    
     @frappe.whitelist()
     def test_connection(self):
         """Test the connection to the integration"""
