+++ doc2sys/integrations/connectors/quickbooks.py
@@ -5,7 +5,6 @@ import json
 
 from doc2sys.integrations.base import BaseIntegration
 from doc2sys.integrations.registry import register_integration
-from doc2sys.integrations.utils import map_document_fields
 
 @register_integration
 class QuickBooksIntegration(BaseIntegration):
